rename inlineStyles to stylesConstants
